requirements.txt
openai
python-telegram-bot
deepgram-sdk
requests
python-dotenv
flask
psycopg2-binary
